requirements.txt
django
django-notifications-hq
django-filter
Django-Verify-Email
django-sslserver